Release diesel 1.4.0
New Features
==

In contrast to the last release most changes in this release are minor or internal. We've added support for newer versions of some dependency crates (libsqlite-sys, uuid, ipnetwork).

Diesel CLI got a command line flag to check if a generated schema matches  the already existing one. This is useful for CI setups to check there if the committed generated schema file matches the committed migrations.

We've added support for the `diesel_mange_updated_at('table_name')` SQL function on SQLite. This function handles the setup of an trigger that automatically updates the `updated_at` column on update operations.

Additionally several helpers were added to support the deserialisation of tuples and composite types on PostgreSQL.

As always, for a full list of changes you can find it in [the changelog](https://github.com/diesel-rs/diesel/blob/v1.4.0/CHANGELOG.md)

Thanks
==

Thank you to everyone who helped make this release happen through bug reports, and discussion on Gitter. While we don't have a way to collect stats on that form of contribution, it's greatly appreciated.

In addition to the Diesel core team, 29 people contributed code to this release. A huge thank you to:

* Aleksey Ivanov
* Andrew Speed
* Carlos Diaz-Padron
* Daniel Silverstone
* Dirkjan Ochtman
* Dominik Sander
* Eyal Kalderon
* Galuh Sahid
* Hirokazu Hata
* Jake Goulding
* Jo Liss
* Josh Leeb-du Toit
* Kevin Stenerson
* kpcyrd
* Marcus Stollsteimer
* Matej Stuchlik
* Nikita Sivakov
* notryanb
* Rasmus Kaj
* Richard Petrie
* Rotem Yaari
* Ryan Leckey
* Sergio Benitez
* Simon Heath
* Stephen Muss
* Trinh Hoang Anh